requirements.txt
flask
sqlalchemy
flask_sqlalchemy
flask_cors
flask_marshmallow
sqlalchemy_utils
marshmallow-sqlalchemy
mysqlclient
pyjwt
winapps
wmi
Flask-APScheduler